Refactoring of SecOC Layer and implementation for SecOC over CANFD
